backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Cors==4.0.0
GeoAlchemy2==0.14.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
